segmentation_python/requirements.txt
appdirs==1.4.3
backports.weakref==1.0rc1
bleach==1.5.0
certifi==2017.7.27.1
chardet==3.0.4
cycler==0.10.0
decorator==4.1.2
entrypoints==0.2.3
html5lib==0.9999999
idna==2.6
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
ipywidgets==7.0.0
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.1.0
jupyter-core==4.3.0
Markdown==2.2.0
MarkupSafe==1.0
matplotlib==2.0.2
mistune==0.7.4
nbconvert==5.2.1
nbformat==4.3.0
notebook==5.0.0
numpy==1.13.0
olefile==0.44
packaging==16.8
pandas==0.20.3
pandocfilters==1.4.2
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.1
progressbar2==3.20.0
prompt-toolkit==1.0.15
protobuf==3.3.0
ptyprocess==0.5.2
pyglet==1.2.4
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
python-utils==2.1.0
pytz==2017.2
pyzmq==16.0.2
qtconsole==4.3.1
requests==2.18.4
scipy==0.19.0
seaborn==0.8
simplegeneric==0.8.1
six==1.10.0
tensorflow-gpu==1.2.0
terminado==0.6
testpath==0.3.1
Theano==0.9.0
tornado==4.5.1
traitlets==4.3.2
url==0.4.1
urllib3==1.22
urllib5==5.0.0
wcwidth==0.1.7
Werkzeug==0.12.2
widgetsnbextension==3.0.0
